Ryddet i script.js
ryddet opp i script.js, la til mapFunction.js fil og endret på meny css.
