+++ script.js
@@ -5,6 +5,16 @@ import { fetchGeoJSONFot, fetchGeoJSONSki, fetchGeoJSONSykkel, fetchGeoJSONAnnen
 import { fetchGeoJSONSkredFaresone } from './skredFaresone.js';
 import { fetchGeoJSONKvikkleireFare } from './kvikkleireFare.js';
 import { nveBratthetLayer, createBratthetLegend } from './bratthet.js';
+import { 
+    enableDynamicLoading, 
+    addLayerToggleButtons, 
+    setupMenuToggle,
+    enableGeolocation, 
+    setStartPosition, 
+    setEndPosition, 
+    calculateDistance, 
+    updateRouteWithUserAddresses 
+} from './mapFunctions.js';
 
 // Supabase URL og API-nøkkel
 const supabaseUrl = 'https://bpttsywlhshivfsyswvz.supabase.co';
@@ -18,7 +28,7 @@ L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png').addTo(map);
 // Felles funksjon for å opprette lag
 const createLayer = () => L.layerGroup();
 
-
+// Definer lagene
 const layers = {
     routeInfo: { layer: createLayer(), visible: false, fetchFunction: fetchGeoJSONRuteInfo },
     route: { layer: createLayer(), visible: false, fetchFunction: fetchGeoJSONAnnen },
@@ -31,214 +41,36 @@ const layers = {
     nveBratthet: nveBratthetLayer
 };
 
-let legend; 
+// Aktiver dynamisk lasting av lag
+enableDynamicLoading(map, layers);
 
-const toggleLayer = async (id) => {
-    const layerConfig = layers[id];
-    if (!layerConfig) {
-        console.error(`Lagkonfigurasjon for ${id} mangler.`);
-        return;
-    }
+// Legg til knapper for å vise/skjule lag
+addLayerToggleButtons(map, layers);
 
-    const { layer, visible } = layerConfig;
-    const button = document.getElementById(`show${capitalizeFirstLetter(id)}`);
+// Aktiver menyknappen
+setupMenuToggle();
 
-    if (visible) {
-        map.removeLayer(layer);
-        button.textContent = `Vis ${capitalizeFirstLetter(id)}`;
-        layerConfig.visible = false;
+// Aktiver geolokalisering
+enableGeolocation(map);
 
-        if (legend) {
-            map.removeControl(legend);
-            legend = null;
-        }
-    } else {
-        layer.addTo(map);
-        button.textContent = `Skjul ${capitalizeFirstLetter(id)}`;
-        layerConfig.visible = true;
-
-        if (id === 'nveBratthet') {
-            legend = createBratthetLegend();
-            legend.addTo(map);
-        }
-    }
-};
-
-// Dynamisk lasting ved flytting eller zooming
-map.on('moveend', async () => {
-    Object.keys(layers).forEach(async (id) => {
-        const layerConfig = layers[id];
-        if (layerConfig.visible && layerConfig.fetchFunction) {
-            // Hent data dynamisk for synlig utsnitt hvis fetchFunction er definert
-            await layerConfig.fetchFunction(map, layerConfig.layer);
-        }
-    });
-});
-// Funksjon for å kapitalisere første bokstav i en streng
-const capitalizeFirstLetter = (string) => string.charAt(0).toUpperCase() + string.slice(1);
-
-// Legg til hendelser for alle knapper
-['routeInfo', 'route', 'hytter', 'fotRuter', 'skiloyper', 'sykkelruter', 'skredFaresone', 'kvikkleireFare', 'nveBratthet'].forEach((id) => {
-    const button = document.getElementById(`show${capitalizeFirstLetter(id)}`);
-    if (button) {
-        button.addEventListener('click', () => toggleLayer(id));
-    } else {
-        console.warn(`Knappen med ID show${capitalizeFirstLetter(id)} finnes ikke.`);
-    }
-});
-
-
-
-const toggleMenuButton = document.getElementById('toggleMenu');
-const menuContent = document.getElementById('menuContent');
-
-toggleMenuButton.addEventListener('click', () => {
-    if (menuContent.style.display === 'none' || menuContent.style.display === '') {
-        menuContent.style.display = 'block'; // Vis menyen
-    } else {
-        menuContent.style.display = 'none'; // Skjul menyen
-    }
-});
-
-// Legg til geolokalisering for å finne brukerens posisjon
-map.locate({ setView: true, maxZoom: 16 });
-
-// Håndtering av når brukerens posisjon er funnet
-map.on('locationfound', (e) => {
-    const userMarker = L.marker(e.latlng).addTo(map);
-    userMarker.bindPopup("Du er her!").openPopup();
-    map.setView(e.latlng, 16);
-});
-
-// Håndtering av feil når posisjonen ikke kan finnes
-map.on('locationerror', (e) => {
-    alert("Kunne ikke finne din posisjon: " + e.message);
-
-});
-
+// Håndter klikk for å sette start- og sluttposisjon
 let startMarker = null;
 let endMarker = null;
 
-// Funksjon for å legge til en markør
-const addMarker = (latlng, label, color) => {
-    return L.marker(latlng, {
-        icon: L.divIcon({
-            className: 'custom-marker',
-            html: `<div style="background-color: ${color}; width: 20px; height: 20px; border-radius: 50%;"></div>`,
-            iconSize: [20, 20],
-            iconAnchor: [10, 10]
-        })
-    }).bindPopup(label).addTo(map);
-};
-
-// Funksjon for å sette startposisjon
-const setStartPosition = (latlng) => {
-    if (startMarker) {
-        map.removeLayer(startMarker);
-    }
-    startMarker = addMarker(latlng, 'Startposisjon', 'green');
-};
-
-// Funksjon for å sette sluttposisjon
-const setEndPosition = (latlng) => {
-    if (endMarker) {
-        map.removeLayer(endMarker);
-    }
-    endMarker = addMarker(latlng, 'Sluttposisjon', 'red');
-};
-
-// Funksjon for å beregne avstand mellom start og slutt
-const calculateDistance = () => {
-    if (startMarker && endMarker) {
-        const startLatLng = startMarker.getLatLng();
-        const endLatLng = endMarker.getLatLng();
-        const distance = startLatLng.distanceTo(endLatLng); // Avstand i meter
-        alert(`Avstanden mellom start og slutt er ${(distance / 1000).toFixed(2)} km.`);
-    } else {
-        alert('Både start- og sluttposisjon må være satt.');
-    }
-};
-
-// Legg til knapper for å sette start- og sluttposisjon
 map.on('click', (e) => {
-    const { lat, lng } = e.latlng;
     const action = prompt('Vil du sette startposisjon eller sluttposisjon? (skriv "start" eller "slutt")');
     if (action === 'start') {
-        setStartPosition(e.latlng);
+        startMarker = setStartPosition(map, startMarker, e.latlng);
     } else if (action === 'slutt') {
-        setEndPosition(e.latlng);
+        endMarker = setEndPosition(map, endMarker, e.latlng);
     }
 });
 
 // Legg til en knapp for å beregne avstand
 const calculateDistanceButton = document.getElementById('calculateDistance');
 if (calculateDistanceButton) {
-    calculateDistanceButton.addEventListener('click', calculateDistance);
-} else {
-    console.warn('Knappen med ID "calculateDistance" finnes ikke.');
+    calculateDistanceButton.addEventListener('click', () => calculateDistance(startMarker, endMarker));
 }
-
-
-// Hent brukerens posisjon og sett den som startpunkt for routing
-map.locate({ setView: true, maxZoom: 16 });
-
-// Funksjon for å oppdatere ruten basert på brukerens adresser
-const updateRouteWithUserAddresses = async () => {
-    // Be brukeren om å skrive inn start- og sluttadresser
-    const startAddress = prompt('Skriv inn startadresse:');
-    const endAddress = prompt('Skriv inn sluttadresse:');
-
-    if (!startAddress || !endAddress) {
-        alert('Du må skrive inn både start- og sluttadresser.');
-        return;
-    }
-
-    try {
-        // Geokoding av startadresse
-        const startResponse = await fetch(`https://nominatim.openstreetmap.org/search?format=json&q=${encodeURIComponent(startAddress)}`);
-        const startData = await startResponse.json();
-
-        if (startData.length === 0) {
-            alert('Fant ikke startadressen. Sjekk at den er riktig.');
-            return;
-        }
-
-        const startLat = parseFloat(startData[0].lat);
-        const startLng = parseFloat(startData[0].lon);
-
-        // Geokoding av sluttadresse
-        const endResponse = await fetch(`https://nominatim.openstreetmap.org/search?format=json&q=${encodeURIComponent(endAddress)}`);
-        const endData = await endResponse.json();
-
-        if (endData.length === 0) {
-            alert('Fant ikke sluttadressen. Sjekk at den er riktig.');
-            return;
-        }
-
-        const endLat = parseFloat(endData[0].lat);
-        const endLng = parseFloat(endData[0].lon);
-
-        // Opprett routing med de geokodede punktene
-        L.Routing.control({
-            waypoints: [
-                L.latLng(startLat, startLng), // Startpunkt
-                L.latLng(endLat, endLng)      // Sluttpunkt
-            ],
-            routeWhileDragging: true, // Tillat rutejustering ved dragging
-            geocoder: L.Control.Geocoder.nominatim() // Bruk geokoding for adresser
-        }).addTo(map);
-
-        // Marker start- og sluttpunktene
-        const startMarker = L.marker([startLat, startLng]).addTo(map);
-        startMarker.bindPopup(`Startpunkt: ${startAddress}`).openPopup();
-
-        const endMarker = L.marker([endLat, endLng]).addTo(map);
-        endMarker.bindPopup(`Sluttpunkt: ${endAddress}`).openPopup();
-    } catch (error) {
-        console.error('Feil under geokoding:', error);
-        alert('Det oppsto en feil under geokoding. Prøv igjen senere.');
-    }
-};
 
 // Legg til en knapp for å oppdatere ruten med brukerens adresser
 const userInputRouteButton = document.createElement('button');
@@ -249,4 +81,4 @@ userInputRouteButton.style.left = '10px';
 userInputRouteButton.style.zIndex = '1000';
 document.body.appendChild(userInputRouteButton);
 
-userInputRouteButton.addEventListener('click', updateRouteWithUserAddresses);
+userInputRouteButton.addEventListener('click', () => updateRouteWithUserAddresses(map));
